pyproject.toml
[tool.poetry]
name = "security"
version = "0.1.0"
description = "some practice on known solidity security issuess"
authors = ["Tang Wei <wei.tang@coinflex.com>"]

[tool.poetry.dependencies]
python = "3.8"
eth-brownie = "^1.17.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
